moved PastVisitsPage to separate file
